ai_platform_engineering/agents/webex/pyproject.toml
[project]
name = "agent_webex"
version = "0.1.0"
license = "Apache-2.0"
description = "Webex natural language agent using LangChain, LangGraph and MCP."
readme = "README.md"
authors = [
    { name = "Torben Tretau", email = "ttretau@cisco.com" },
]
maintainers = [
    { name = "Torben Tretau", email = "ttretau@cisco.com" },
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "a2a-sdk==0.2.16",
    "agentevals>=0.0.7",
    "agntcy-app-sdk==0.1.4",
    "click>=8.2.0",
    "langchain-anthropic>=0.3.13",
    "langchain-core>=0.3.60",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph==0.5.3",
    "pytest>=8.3.5",
    "tabulate>=0.9.0",
    "uv",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.2",
    "httpx>=0.27.0",
    "httpx-sse>=0.3.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "rich (>=14.0.0,<15.0.0)",
    "sseclient (>=0.0.27,<0.0.28)",
    "cnoe-agent-utils==0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.poetry.scripts]
agent_webex_a2a = "agent_webex.protocol_bindings.a2a_server:main"

[tool.poetry.dependencies]
uv = "^0.7.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.11",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
]

[tool.ruff]
# Allow lines to be as long as 140.
line-length = 140
# Number of spaces per indentation level
indent-width = 2

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
]
ignore = ["F403"]
